# swagger/config.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.0",
    "title": "Invoices APIs Document",
    "description": "APIs document for all operations on the inboxes app.",
    "termsOfService": "",
    "contact": {
      "name": "Phathutshedzo Khabubu"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
  "tags": [
    {
      "name": "Invoices",
      "description": "API for invoices in the system"
    }
  ],
  "schemes": ["http"],
  "paths": {
    "/invoices": {
      "get": {
        "tags": ["invoices"],
        "summary": "Get all invoices",
        "description": "Get all invoices",
        "operationId": "getInvoices",
        "responses": {
          "200": {
            "description": "Sucessful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Invoice"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": "false"
                    },
                    "error": {
                      "type": "string",
                      "example": "Something wrong happened please try again later!"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["invoices"],
        "summary": "Generate an invoice",
        "description": "Generate an invoice",
        "operationId": "postInvoice",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Invoice"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Invoice is created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": "false"
                    },
                    "error": {
                      "type": "string",
                      "example": "Missing fields!!"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": "false"
                    },
                    "error": {
                      "type": "string",
                      "example": "Something wrong happened please try again later!"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["invoices"],
        "summary": "Update an invoice",
        "description": "Update an invoice",
        "operationId": "updateInvoice",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Invoice"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invoice is updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": "false"
                    },
                    "error": {
                      "type": "string",
                      "example": "Something wrong happened please try again later!"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Specifies resource does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": "false"
                    },
                    "error": {
                      "type": "string",
                      "example": "The specified invoice is not found!"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": "false"
                    },
                    "error": {
                      "type": "string",
                      "example": "Something wrong happened please try again later!"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/invoices/:id": {
      "get": {
        "tags": ["invoices"],
        "summary": "Get an invoice",
        "description": "Get an invoice",
        "operationId": "getInvoice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of invoice that needs to be fetched",
            "required": true,
            "schema": {
              "type": "string",
              "example": "62d95227a7c11a49cdd3b475"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operation is successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "404": {
            "description": "Operation is unsuccessful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": "false"
                    },
                    "error": {
                      "type": "string",
                      "example": "The invoice does not exist"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": "false"
                    },
                    "error": {
                      "type": "string",
                      "example": "Something wrong happened please try again later!"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["invoices"],
        "summary": "Delete an invoice",
        "description": "Delete an invoice",
        "operationId": "deleteInvoice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of invoice that needs to be deleted",
            "required": true,
            "schema": {
              "type": "string",
              "example": "62d95227a7c11a49cdd3b475"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice is deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": "true"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Operation is unsuccessful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": "false"
                    },
                    "error": {
                      "type": "string",
                      "example": "The invoice does not exist"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": "false"
                    },
                    "error": {
                      "type": "string",
                      "example": "Something wrong happened please try again later!"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Invoice": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "format": "int64",
            "example": "62d95227a7c11a49cdd3b475"
          },
          "amount": {
            "type": "number",
            "format": "float",
            "example": 450.99
          },
          "invoiceNumber": {
            "type": "integer",
            "format": "int32",
            "example": 1000889
          },
          "invoiceDate": {
            "type": "string",
            "format": "date"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "xml": {
          "name": "order"
        }
      }
    },
    "requestBodies": {},
    "securitySchemes": {}
  }
}